Pointillism
This program consists of 2 codes; art maker and art viewer.
Art maker first checks to see if the file selected has .gif or not. If it does, it will exit, otherwise it will create a .art file where it will store a random x, random
y, a random radius (ranging from 2-9) and it finds the pixel of the random x, y chosen. This is executed 5000 times.
Art viewer creates the Pointillism. It uses the .art file and creates circles of random pixels created by the art_maker. This program reads every line of the file created by the art_maker and identifies x, y coordinates, color and radius 5000 times (because there are 5000 lines). Also, program this adjusts to the width size if the size of the image is smaller. This program also exits if the file selected doesn't have .art